HAWQ-969. Modify some functions in hdfs_config.cpp and yarn_config.cpp

+++ src/test/feature/lib/hdfs_config.cpp
@@ -14,6 +14,58 @@ using std::string;
 namespace hawq {
 namespace test {
 
+void HdfsConfig::runCommand(const string &command, 
+                            bool ishdfsuser, 
+                            string &result) {
+  string cmd = "";
+  if (ishdfsuser) {
+    cmd = "sudo -u ";
+    cmd.append(getHdfsUser());
+    cmd.append(" ");
+    cmd.append(command);
+  } else {
+    cmd = command;
+  }
+  Command c(cmd);
+  result = c.run().getResultOutput();
+}
+
+bool HdfsConfig::runCommandAndFind(const string &command, 
+                                   bool ishdfsuser, 
+                                   const string &findstring) {
+  string result = "";
+  runCommand(command, ishdfsuser, result);
+  auto lines = hawq::test::split(result, '\n');
+  for (size_t i=0; i<lines.size(); i++) {
+    string valueLine = lines[i];
+    int find = valueLine.find(findstring);
+    if (find >= 0) {
+        return true;
+    }
+  }
+  return false;
+}
+
+void HdfsConfig::runCommandAndGetNodesPorts(const string &command, 
+                                            std::vector<string> &datanodelist,
+                                            std::vector<int> &port) {
+  string result = "";
+  runCommand(command, true, result);
+  auto lines = hawq::test::split(result, '\n');
+  for (size_t i = 0; i < lines.size(); i++) {
+    string valueLine = lines[i];
+    auto datanodeInfo = hawq::test::split(valueLine, ':');
+    if (datanodeInfo.size() == 3) {
+      int portStart = datanodeInfo[2].find_first_of('(');
+      int portEnd = datanodeInfo[2].find_first_of(')');
+      string datanodePort = datanodeInfo[2].substr(0, portStart);
+      string datanodeHost = datanodeInfo[2].substr(portStart+1, portEnd-portStart-1);
+      datanodelist.push_back(hawq::test::trim(datanodeHost));
+      port.push_back(std::stoi(hawq::test::trim(datanodePort)));
+    }
+  }
+}
+
 string HdfsConfig::getHdfsUser() {
   string cmd = "ps aux|grep hdfs.server|grep -v grep";
   Command c(cmd);
@@ -26,6 +78,9 @@ string HdfsConfig::getHdfsUser() {
 }
 
 bool HdfsConfig::LoadFromHawqConfigFile() {
+  if (isLoadFromHawqConfigFile) {
+    return true;
+  }
   const char *env = getenv("GPHOME");
   string confPath = env ? env : "";
   if (confPath != "") {
@@ -35,22 +90,31 @@ bool HdfsConfig::LoadFromHawqConfigFile() {
   }
 
   hawqxmlconf.reset(new XmlConfig(confPath));
-  hawqxmlconf->parse();
+  if (!hawqxmlconf->parse())
+    return false;
+
+  isLoadFromHawqConfigFile = true;
   return true;
 }
 
 bool HdfsConfig::LoadFromHdfsConfigFile() {
+  if (isLoadFromHdfsConfigFile) {
+    return true;
+  }
   string confPath=getHadoopHome();
   if (confPath == "") {
     return false;
   }
   confPath.append("/etc/hadoop/hdfs-site.xml");
   hdfsxmlconf.reset(new XmlConfig(confPath));
-  hdfsxmlconf->parse();
+  if (!hdfsxmlconf->parse())
+    return false;
+  
+  isLoadFromHdfsConfigFile = true;
   return true;
 }
 
-bool HdfsConfig::isHA() {
+int HdfsConfig::isHA() {
   const hawq::test::PSQLQueryResult &result = psql.getQueryResult(
        "SELECT substring(fselocation from length('hdfs:// ') for (position('/' in substring(fselocation from length('hdfs:// ')))-1)::int) "
        "FROM pg_filespace pgfs, pg_filespace_entry pgfse "
@@ -59,53 +123,46 @@ bool HdfsConfig::isHA() {
   if (table.size() > 0) {
     int find = table[0][0].find(":");
     if (find < 0) {
-      return true;
+      return 1;
     } else {
-      return false;
+      return 0;
     }
   }
-  return false;
+  return -1;
 }
 
-bool HdfsConfig::isConfigKerberos() {
+int HdfsConfig::isConfigKerberos() {
   bool ret = LoadFromHawqConfigFile();
   if (!ret) {
-    throw GetHawqHomeException();
+    return -1;
   }
   string authentication = hawqxmlconf->getString("hadoop.security.authentication");
   if (authentication == "kerberos") {
-    return true;
+    return 1;
   } else {
-    return false;
+    return 0;
   }
 }
 
-bool HdfsConfig::isTruncate() {
-  string cmd = "hadoop fs -truncate";
-  Command c(cmd);
-  string result = c.run().getResultOutput();
-  auto lines = hawq::test::split(result, '\n');
-  if (lines.size() >= 1) {
-    string valueLine = lines[0];
-    int find = valueLine.find("-truncate: Unknown command");
-    if (find < 0) {
-      return true;
-    }
+int HdfsConfig::isTruncate() {
+  if (runCommandAndFind("hadoop fs -truncate", false, "-truncate: Unknown command")) {
+    return 0;
+  } else {
+    return 1;
   }
-  return false;
 }
 
 string HdfsConfig::getHadoopHome() {
-  string cmd = "ps -ef|grep hadoop";
-  Command c(cmd);
-  string result = c.run().getResultOutput();
+  string result = "";
+  runCommand("ps -ef|grep hadoop", true, result);
   string hadoopHome = "";
   auto lines = hawq::test::split(result, '\n');
   for (size_t i=0; i<lines.size()-1; i++) {
     string valueLine = lines[i];
-    int pos = valueLine.find("-Dhadoop.home.dir=");
+    string findstring = "-Dhadoop.home.dir=";
+    int pos = valueLine.find(findstring);
     if (pos >=0 ) {
-      string valueTmp = valueLine.substr(pos+18); 
+      string valueTmp = valueLine.substr(pos+findstring.size()); 
       int valueEnd = valueTmp.find_first_of(" ");
       string value = valueTmp.substr(0, valueEnd);
       hadoopHome = hawq::test::trim(value);
@@ -139,19 +196,11 @@ bool HdfsConfig::getHANamenode(const string &namenodetype,
   auto haNamenodes = hawq::test::split(haNamenodesValue, ',');
   for (size_t i = 0; i < haNamenodes.size(); i++) {
     string haNamenode = hawq::test::trim(haNamenodes[i]);
-    string cmd = "sudo -u ";
-    cmd.append(getHdfsUser());
-    cmd.append(" hdfs haadmin -getServiceState ");
+    string cmd = "hdfs haadmin -getServiceState ";
     cmd.append(haNamenode);
-    Command c(cmd);
-    string result = c.run().getResultOutput();
-    auto lines = hawq::test::split(result, '\n');
-    if (lines.size() >= 1) {
-      string valueLine = lines[0];
-      if (valueLine == namenodetype) {
-        namenodeService = haNamenode;
-        break;
-      }
+    if (runCommandAndFind(cmd, true, namenodetype)) {
+      namenodeService = haNamenode;
+      break;
     }
   }
   string rpcAddressName = "dfs.namenode.rpc-address.gphd-cluster.";
@@ -165,11 +214,8 @@ bool HdfsConfig::getHANamenode(const string &namenodetype,
 
 void HdfsConfig::getNamenodes(std::vector<string> &namenodes,
                               std::vector<int> &port) {
-  string cmd = "sudo -u ";
-  cmd.append(getHdfsUser());
-  cmd.append(" hdfs getconf -nnRpcAddresses");
-  Command c(cmd);
-  string result = c.run().getResultOutput();
+  string result = "";
+  runCommand("hdfs getconf -nnRpcAddresses", true, result);
   auto lines = hawq::test::split(result, '\n');
   for (size_t i = 0; i < lines.size(); i++) {
     string valueLine = lines[i];
@@ -183,71 +229,28 @@ void HdfsConfig::getNamenodes(std::vector<string> &namenodes,
 
 void HdfsConfig::getDatanodelist(std::vector<string> &datanodelist,
                                  std::vector<int> &port) {
-  string cmd = "sudo -u ";
-  cmd.append(getHdfsUser());
-  cmd.append(" hdfs dfsadmin -report | grep Name");
-  Command c(cmd);
-  string result = c.run().getResultOutput();
-  auto lines = hawq::test::split(result, '\n');
-  for (size_t i = 0; i < lines.size(); i++) {
-    string valueLine = lines[i];
-    auto datanodeInfo = hawq::test::split(valueLine, ':');
-    if (datanodeInfo.size() == 3) {
-      int portStart = datanodeInfo[2].find_first_of('(');
-      int portEnd = datanodeInfo[2].find_first_of(')');
-      string datanodePort = datanodeInfo[2].substr(0, portStart);
-      string datanodeHost = datanodeInfo[2].substr(portStart+1, portEnd-portStart-1);
-      datanodelist.push_back(hawq::test::trim(datanodeHost));
-      port.push_back(std::stoi(hawq::test::trim(datanodePort)));
-    }
-  }
+  runCommandAndGetNodesPorts("hdfs dfsadmin -report | grep Name", datanodelist, port);
 }
 
 void HdfsConfig::getActiveDatanodes(std::vector<string> &activedatanodes,
                                     std::vector<int> &port) {
-  string cmd = "sudo -u ";
-  cmd.append(getHdfsUser());
-  cmd.append(" hdfs dfsadmin -report -live | grep Name");
-  Command c(cmd);
-  string result = c.run().getResultOutput();
-  auto lines = hawq::test::split(result, '\n');
-  for (size_t i = 0; i < lines.size(); i++) {
-    string valueLine = lines[i];
-    auto datanodeInfo = hawq::test::split(valueLine, ':');
-    if (datanodeInfo.size() == 3) {
-      int portStart = datanodeInfo[2].find_first_of('(');
-      int portEnd = datanodeInfo[2].find_first_of(')');
-      string datanodePort = datanodeInfo[2].substr(0, portStart);
-      string datanodeHost = datanodeInfo[2].substr(portStart+1, portEnd-portStart-1);
-      activedatanodes.push_back(hawq::test::trim(datanodeHost));
-      port.push_back(std::stoi(hawq::test::trim(datanodePort)));
-    }
-  }
+  runCommandAndGetNodesPorts("hdfs dfsadmin -report -live | grep Name", activedatanodes, port);
 }
 
-
-bool HdfsConfig::isSafemode() {
-  string cmd = "hadoop fs -mkdir /tmp_hawq_test";
-  Command c(cmd);
-  string result = c.run().getResultOutput();
-  auto lines = hawq::test::split(result, '\n');
-  if (lines.size() >= 1) {
-    string valueLine = lines[0];
-    int find = valueLine.find("Name node is in safe mode.");
-    if (find >= 0) {
-      return true;
-    }
+int HdfsConfig::isSafemode() {
+  if (runCommandAndFind("hadoop fs -mkdir /tmp_hawq_test", false, "Name node is in safe mode.")) {
+    return 1;
   }
-  cmd = "hadoop fs -rm -r /tmp_hawq_test";
+  string cmd = "hadoop fs -rm -r /tmp_hawq_test";
   Command c_teardown(cmd);
-  result = c_teardown.run().getResultOutput();
-  return false;
+  string result = c_teardown.run().getResultOutput();
+  return 0;
 }
 
 string HdfsConfig::getParameterValue(const string &parameterName) {
   bool ret = LoadFromHdfsConfigFile();
   if (!ret) {
-    throw GetHadoopHomeException();
+    return "Error: failed to load from HDFS configuration file";
   }
 
   return hdfsxmlconf->getString(parameterName);
@@ -255,11 +258,12 @@ string HdfsConfig::getParameterValue(const string &parameterName) {
 
 string HdfsConfig::getParameterValue(const string &parameterName,
                                      const string &conftype) {
-  if (conftype == "hdfs" || conftype == "HDFS")
+  if (hawq::test::lower(conftype) == "hdfs") {
     return getParameterValue(parameterName);
+  }
   bool ret = LoadFromHawqConfigFile();
   if (!ret) {
-    throw GetHawqHomeException();
+    return "Error: failed to load from HAWQ configuration file";
   }
 
   return hawqxmlconf->getString(parameterName);
@@ -269,7 +273,7 @@ bool HdfsConfig::setParameterValue(const string &parameterName,
                                    const string &parameterValue) { 
   bool ret = LoadFromHdfsConfigFile();
   if (!ret) {
-    throw GetHadoopHomeException();
+    return false;
   }
 
   return hdfsxmlconf->setString(parameterName, parameterValue);
@@ -278,12 +282,12 @@ bool HdfsConfig::setParameterValue(const string &parameterName,
 bool HdfsConfig::setParameterValue(const string &parameterName,
                                    const string &parameterValue,
                                    const string &conftype) {
-  if (conftype == "hdfs" || conftype == "HDFS") {
+  if (hawq::test::lower(conftype) == "hdfs") {
     return setParameterValue(parameterName, parameterValue);
   }
   bool ret = LoadFromHawqConfigFile();
   if (!ret) {
-    throw GetHawqHomeException();
+    return false;
   }
 
   return hawqxmlconf->setString(parameterName, parameterValue);
